+++ paymentmethod/HeidelpaySEPADirectDebitGuaranteed.php
@@ -52,6 +52,16 @@ class HeidelpaySEPADirectDebitGuaranteed extends HeidelpayPaymentMethod implemen
     use TranslatorTrait;
 
     /**
+     * Deactivate as the payment method is deprecated
+     * @param array $args
+     * @return bool
+     */
+    public function isValidIntern($args = []): bool
+    {
+        return false;
+    }
+
+    /**
      * Cancel the Charge.
      *
      * Invoice factoring has an additional mandatory field (reason code) in case of a cancel.